Create Log count verifying alert
Search string that uses a hourly generated file to count the logs coming into Splunk and find the abnormalities to alert on. This search is great for finding any unexpected increase or decrease to log data and should run mostly universally on any environment though some tweaking and a .csv lookup is required  To a future Skyphex that might be dumb enough to generate this on his own environment the format for the .csv file is: index, source, hourly_date, Event_Count.