Create Self Email Search
This report would need a heavy overhaul to work on other environments but has several string comparison lines that i built using regex that spunk has no other option to use unless an app is involved making parts of this perfect for environments with minimal setup or the inability to download apps. This search specifically takes emails that are being sent out and uses Active Directory data as well as name and email data to determine if a user is emailing something to a personal email. The severity levels and determining factors will vary greatly from environment to environment